Make code utils a public API
